fixed users controller
Former-commit-id: 1dd5c1ed5d4dae6d6ea3297c308a2ae60e3ff795

+++ app/Http/Controllers/UsersController.php
@@ -257,7 +257,7 @@ class UsersController extends Controller
             $delegation = UserDelegation::find($id_ud->id);
             $delegation ->delegations_id =$request->delegation_id;
             $delegation->update();
-        }else{
+        }elseif($id_ud != null && $request->delegation_id==null){
             $id_ud=UserDelegation::where('users_id','=',$id)->first();
             $id_ud->delete();
         }
